fix analyzer diagnostics in encoder bounds and byte offsets

+++ examples/dstar_decode.c
@@ -57,7 +57,7 @@ write_samples(FILE* fp, short pcm[FRAME_SAMPLES], const char bits[49]) {
             value = -31128;
         }
         uint16_t sample = (uint16_t)value;
-        bytes[2 * i] = (unsigned char)sample;
+        bytes[2 * (size_t)i] = (unsigned char)sample;
         bytes[2 * i + 1] = (unsigned char)(sample >> 8);
     }
     return fwrite(bytes, 1, sizeof(bytes), fp) == sizeof(bytes) ? 0 : -1;
